Remove all trianglify functionality; add static trianglify background image
Screw it! The way that trianglify's image generation overflows its containers no matter what is not worth the trouble of having a dynamic decorative element. I went to the creator's website and used the image generator to create a static image, which will now be the background. Maybe I'll generate a few images and dynamically pick a random one, but I've spent too much time on the UI. I need to work on the back-end.